backend/pyproject.toml
[tool.poetry]
name = "concentrate-qual-backend"
version = "1.0.0"
description = "Fixing and viewing concentrate values monthly"
authors = ["KurushkinDA <kurushkin.dimka@yandex.ru>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
pydantic = "^2.11.4"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"